Extract DelegatedTransformer
